Update and rename adolf_pk.js to ddo_pk.js
